deal with signedness annoyances

+++ volume/dsp_volume.cpp
@@ -14,7 +14,7 @@ void DspVolume::process(gsl::span<int16_t> samples, int32_t channels)
     do_process(samples);
 }
 
-float DspVolume::fade_step(int32_t frame_count)
+float DspVolume::fade_step(size_t frame_count)
 {
     // compute manual gain
     auto current_gain = gain_current();
